#286 dont cast fps to int

+++ code/src/main/python/model/data/timeline.py
@@ -66,7 +66,7 @@ class TimelineModel:
         return self.timeline
 
     def get_fps(self):
-        return int(self.timeline.info.fps.num / self.timeline.info.fps.den)
+        return self.timeline.info.fps.num / self.timeline.info.fps.den
 
     def change(self, change_type, key, data):
         """
